feat(cli): add configurable I/O and queue depth for stress workloads
Rollback trigger: test failures or integration instability
